change mode to maintain all secrets files in ~/.dotfiles checkout
 - still are not maintained in git (excluded in top-level .gitignore)
 - one directory hierarchy contains all the secrets

To use, add a symlink here from desired filename to .secrets file. Example:

```
ln -s api-key.secrets.txt api-key.txt
```

The symlink just created (`api-key.txt` in this example) is what stow will
point to and what is committed to git repo. It is committed as a symlink type in the git repo.